endpoints protegidos por token y roles

+++ backend/controllers/comentarioController.php
@@ -1,5 +1,7 @@
 <?php
     require_once __DIR__ . '/../models/Comentario.php';
+    require_once __DIR__ . '/../helpers/auth.php';
+
     class ComentarioController {
         public function index() {
             $comentario = new Comentario();
@@ -14,10 +16,24 @@
         }
         public function delete($id) {
             $comentario = new Comentario();
+            $comentarioDatos = $comentario->get($id);
+            $idUsuario = $comentarioDatos->usuario_id;
+            $payload = verificarTokenYRol("admin", "usuario");
+            if (!$payload || $payload->sub != $idUsuario) {
+                http_response_code(403);
+                echo json_encode(["message" => "Acceso denegado"]);
+                return;
+            }
             $comentario->delete($id);
             echo json_encode(["message" => "Comentario eliminado"]);
         }
         public function store(){
+            $payload = verificarTokenYRol("admin", "usuario");
+            if (!$payload) {
+                http_response_code(403);
+                echo json_encode(["message" => "Acceso denegado"]);
+                return;
+            }
              $json = file_get_contents('php://input');
              $data = json_decode($json, true);
              $comentario = new Comentario($data['usuario_id'], $data['review_id'], $data['contenido'], $data['fecha']);
@@ -26,6 +42,15 @@
         }
 
         public function update($id){
+            $comentario = new Comentario();
+            $comentarioDatos = $comentario->get($id);
+            $idUsuario = $comentarioDatos->usuario_id;
+            $payload = verificarTokenYRol("admin", "usuario");
+            if (!$payload || $payload->sub != $idUsuario) {
+                http_response_code(403);
+                echo json_encode(["message" => "Acceso denegado"]);
+                return;
+            }
             $json = file_get_contents('php://input');
             $data = json_decode($json, true);
             $comentario = new Comentario($data['usuario_id'], $data['review_id'], $data['contenido'], $data['fecha']);
@@ -35,6 +60,14 @@
 
         public function destroy($id){
             $comentario = new Comentario();
+            $comentarioDatos = $comentario->get($id);
+            $idUsuario = $comentarioDatos->usuario_id;
+            $payload = verificarTokenYRol("admin", "usuario");
+            if (!$payload || $payload->sub != $idUsuario) {
+                http_response_code(403);
+                echo json_encode(["message" => "Acceso denegado"]);
+                return;
+            }
             $comentario->delete($id);
             echo json_encode(["message" => "Comentario eliminado"]);
         }
