+++ backend/controllers/reviewController.php
@@ -1,5 +1,7 @@
 <?php
     require_once __DIR__ . '/../models/Review.php';
+    require_once __DIR__ . '/../helpers/auth.php';
+
     class ReviewController {
         public function index() {
             $review = new Review();
@@ -15,11 +17,27 @@
 
         public function destroy($id) {
             $review = new Review();
+            $reviewDatos = $review->get($id);
+            $idUsuario = $reviewDatos->usuario_id;
+            $payload = verificarTokenYRol("admin","usuario");
+            
+            if(!$payload || $payload->sub != $idUsuario) {
+                http_response_code(403);
+                echo json_encode(["message" => "Acceso denegado"]);
+                return;
+            }
+            
             $review->delete($id);
             echo json_encode(["message" => "Review eliminado"]);
         }
 
         public function store() {
+            $payload = verificarTokenYRol("admin","usuario");
+            if(!$payload) {
+                http_response_code(403);
+                echo json_encode(["message" => "Acceso denegado"]);
+                return;
+            }
             $json = file_get_contents('php://input');
             $data = json_decode($json, true);
             $review = new Review($data['usuario_id'], $data['videojuego_id'], $data['titulo'], $data['contenido'], $data['puntuacion'], $data['fecha']);
@@ -28,6 +46,15 @@
         }
 
         public function update($id) {
+            $review = new Review();
+            $reviewDatos = $review->get($id);
+            $idUsuario = $reviewDatos->usuario_id;
+            $payload = verificarTokenYRol("admin","usuario");
+            if(!$payload || $payload->sub != $idUsuario) {
+                http_response_code(403);
+                echo json_encode(["message" => "Acceso denegado"]);
+                return;
+            }
             $json = file_get_contents('php://input');
             $data = json_decode($json, true);
             $review = new Review($data['usuario_id'], $data['videojuego_id'], $data['titulo'], $data['contenido'], $data['puntuacion'], $data['fecha']);
